+++ packages/apps/graph/src/graph/Query/blocksFromHeight.ts
@@ -1,7 +1,8 @@
 import { prismaClient } from '@db/prismaClient';
+import { COMPLEXITY } from '@services/complexity';
 import { chainIds as defaultChainIds } from '@utils/chains';
 import { normalizeError } from '@utils/errors';
-import { COMPLEXITY, PRISMA, builder } from '../builder';
+import { PRISMA, builder } from '../builder';
 import Block from '../objects/Block';
 
 builder.queryField('blocksFromHeight', (t) => {
