+++ packages/apps/graph/src/graph/Query/transactions.ts
@@ -1,7 +1,11 @@
 import { prismaClient } from '@db/prismaClient';
 import type { Prisma } from '@prisma/client';
+import {
+  COMPLEXITY,
+  getDefaultConnectionComplexity,
+} from '@services/complexity';
 import { normalizeError } from '@utils/errors';
-import { COMPLEXITY, PRISMA, builder } from '../builder';
+import { PRISMA, builder } from '../builder';
 
 builder.queryField('transactions', (t) => {
   return t.prismaConnection({
@@ -16,11 +20,11 @@ builder.queryField('transactions', (t) => {
     type: 'Transaction',
     cursor: 'blockHash_requestKey',
     complexity: (args) => ({
-      field:
-        (args.moduleName
-          ? COMPLEXITY.FIELD.PRISMA_WITH_RELATIONS
-          : COMPLEXITY.FIELD.PRISMA_WITHOUT_RELATIONS) *
-        (args.first || args.last || PRISMA.DEFAULT_SIZE),
+      field: getDefaultConnectionComplexity({
+        withRelations: !!args.moduleName,
+        first: args.first,
+        last: args.last,
+      }),
     }),
     async totalCount(__parent, args) {
       try {
