+++ packages/apps/graph/src/graph/objects/ModuleAccount.ts
@@ -1,9 +1,13 @@
 import { prismaClient } from '@db/prismaClient';
 import { getChainModuleAccount } from '@services/account-service';
+import {
+  COMPLEXITY,
+  getDefaultConnectionComplexity,
+} from '@services/complexity';
 import { chainIds } from '@utils/chains';
 import { dotenv } from '@utils/dotenv';
 import { normalizeError } from '@utils/errors';
-import { COMPLEXITY, PRISMA, builder } from '../builder';
+import { PRISMA, builder } from '../builder';
 import { accountDetailsLoader } from '../data-loaders/account-details';
 import type { ChainModuleAccount, ModuleAccount } from '../types/graphql-types';
 import {
@@ -98,9 +102,11 @@ export default builder.node(
         cursor: 'blockHash_requestKey',
         edgesNullable: false,
         complexity: (args) => ({
-          field:
-            COMPLEXITY.FIELD.PRISMA_WITH_RELATIONS *
-            (args.first || args.last || PRISMA.DEFAULT_SIZE),
+          field: getDefaultConnectionComplexity({
+            withRelations: true,
+            first: args.first,
+            last: args.last,
+          }),
         }),
         async totalCount(parent) {
           try {
@@ -144,9 +150,10 @@ export default builder.node(
         cursor: 'blockHash_chainId_orderIndex_moduleHash_requestKey',
         edgesNullable: false,
         complexity: (args) => ({
-          field:
-            COMPLEXITY.FIELD.PRISMA_WITHOUT_RELATIONS *
-            (args.first || args.last || PRISMA.DEFAULT_SIZE),
+          field: getDefaultConnectionComplexity({
+            first: args.first,
+            last: args.last,
+          }),
         }),
         async totalCount(parent) {
           try {
